move test spec files into test/src folder
